+++ src/bid.rs
@@ -15,8 +15,6 @@ pub(crate) mod score;
 use crate::errors::BlindBidError;
 
 #[cfg(feature = "canon")]
-use canonical::Canon;
-#[cfg(feature = "canon")]
 use canonical_derive::Canon;
 
 use core::borrow::Borrow;
@@ -249,28 +247,28 @@ impl Bid {
     }
 
     /// Returns the `encrypted_data` field of the Bid.
-    pub fn encrypted_data(&self) -> PoseidonCipher {
-        self.encrypted_data
+    pub fn encrypted_data(&self) -> &PoseidonCipher {
+        &self.encrypted_data
     }
 
     /// Returns the `nonce` field of the Bid.
-    pub fn nonce(&self) -> BlsScalar {
-        self.nonce
+    pub fn nonce(&self) -> &BlsScalar {
+        &self.nonce
     }
 
     /// Returns the `hashed_secret` field of the Bid.
-    pub fn hashed_secret(&self) -> BlsScalar {
-        self.hashed_secret
+    pub fn hashed_secret(&self) -> &BlsScalar {
+        &self.hashed_secret
     }
 
     /// Returns the `commitment` field of the Bid.
-    pub fn commitment(&self) -> JubJubAffine {
-        self.c
+    pub fn commitment(&self) -> &JubJubAffine {
+        &self.c
     }
 
     /// Returns the `eligibility` field of the Bid.
-    pub fn eligibility(&self) -> u64 {
-        self.eligibility
+    pub fn eligibility(&self) -> &u64 {
+        &self.eligibility
     }
 
     /// Sets a new value for the eligibility of the Bid.
@@ -279,8 +277,8 @@ impl Bid {
     }
 
     /// Returns the `expiration` field of the Bid.
-    pub fn expiration(&self) -> u64 {
-        self.expiration
+    pub fn expiration(&self) -> &u64 {
+        &self.expiration
     }
 
     /// Returns a mutable ref pointing to the `pos` field of the
@@ -290,8 +288,8 @@ impl Bid {
     }
 
     /// Returns the `pos` field of the Bid.
-    pub fn pos(&self) -> u64 {
-        self.pos
+    pub fn pos(&self) -> &u64 {
+        &self.pos
     }
 
     /// Sets a new value for the position of the Bid.
@@ -389,7 +387,7 @@ mod bid_serialization {
             let stealth_addr = pk_r.gen_stealth_address(&secret);
             let secret = GENERATOR_EXTENDED * secret;
             let value: u64 =
-                (&mut rand::thread_rng()).gen_range(V_RAW_MIN, V_RAW_MAX);
+                (&mut rand::thread_rng()).gen_range(V_RAW_MIN..V_RAW_MAX);
             let value = JubJubScalar::from(value);
             // Set the timestamps as the max values so the proofs do not fail
             // for them (never expired or non-elegible).