+++ CHANGELOG.md
@@ -7,21 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Change error module to include plonk & poseidon err variants [#131](https://github.com/dusk-network/dusk-blindbid/issues/131)
+- Change crate features to not require `std` to use `proof` mod. [#135](https://github.com/dusk-network/dusk-blindbid/issues/135)
+- Change API getters returning by ref and not by value. [#124](https://github.com/dusk-network/dusk-blindbid/issues/124)
+
+### Removed
+- Remove `anyhow` from the crate [#131](https://github.com/dusk-network/dusk-blindbid/issues/131)
+
 ## [0.7.1] - 22-02-21
 
 ### Added
 
-- Added `set_eligibility` for `Bid`. [#121](https://github.com/dusk-network/dusk-blindbid/issues/125)
+- Add `set_eligibility` for `Bid` [#125](https://github.com/dusk-network/dusk-blindbid/issues/125)
 
 ## [0.7.0] - 17-02-21
 
 ### Added
 
-- Added `extend_expiration()` for `Bid`. [#121](https://github.com/dusk-network/dusk-blindbid/issues/121)
+- Add `extend_expiration()` for `Bid` [#121](https://github.com/dusk-network/dusk-blindbid/issues/121)
 
 ### Changed
 
-- Changed `Bid::set_pos()` to not return `&mut u64`. [#121](https://github.com/dusk-network/dusk-blindbid/issues/121)
+- Change `Bid::set_pos()` to not return `&mut u64` [#121](https://github.com/dusk-network/dusk-blindbid/issues/121)
 
 ## [0.6.0] - 12-02-21